auth: also capture the macOS hosted-browser sign-in failure path; split files
Address autoreview:
- The macOS app signs in through HostBrowserSignInFlow (the web app's hosted
  sign-in), which completes via completeExternalSignIn(), not
  AuthCoordinator.signInWithApple(). Capture session-validation failures there
  too (flow "browser_oauth") so the primary macOS OAuth/Apple path is no longer
  invisible. Cancellation is still skipped. Adds a regression test.
- Split AuthErrorReporting.swift into one-type-per-file (AuthErrorReporting,
  AuthErrorContext, NoopAuthErrorReporting) per the package organization rule.
