+++ Packages/CmuxAuthRuntime/Sources/CmuxAuthRuntime/Diagnostics/AuthErrorReporting.swift
@@ -25,37 +25,3 @@ public protocol AuthErrorReporting: Sendable {
     ///     non-sensitive strings safe to send to a telemetry backend.
     func report(error: any Error, context: AuthErrorContext)
 }
-
-/// Structured, non-sensitive context attached to a captured auth failure.
-///
-/// The conforming reporter (which owns the backend SDK) is responsible for
-/// extracting any stable error code from the reported error itself; this
-/// context only carries the flow-level facts the coordinator knows.
-public struct AuthErrorContext: Sendable, Equatable {
-    /// The auth flow that failed (e.g. `"oauth"`, `"password"`, `"magic_link"`).
-    public let flow: String
-    /// The OAuth provider when ``flow`` is `"oauth"` (e.g. `"apple"`), else `nil`.
-    public let provider: String?
-
-    /// Creates an auth error context.
-    /// - Parameters:
-    ///   - flow: The failing auth flow.
-    ///   - provider: The OAuth provider, when applicable.
-    public init(flow: String, provider: String? = nil) {
-        self.flow = flow
-        self.provider = provider
-    }
-}
-
-/// A no-op ``AuthErrorReporting`` used as the default and in tests.
-///
-/// Reporting auth failures is opt-in: an app that does not inject a real
-/// reporter (e.g. a unit-test host, or iOS before its diagnostics reporter is
-/// wired) silently drops the events rather than depending on a backend.
-public struct NoopAuthErrorReporting: AuthErrorReporting {
-    /// Creates a no-op reporter.
-    public init() {}
-
-    /// Discards the reported error.
-    public func report(error: any Error, context: AuthErrorContext) {}
-}
